// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AgentRole = {
  ADMIN: 'ADMIN',
  SUPPORT: 'SUPPORT'
} as const

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole]


export const TicketStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  WAITING_CUSTOMER: 'WAITING_CUSTOMER',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const TicketPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT'
} as const

export type TicketPriority = (typeof TicketPriority)[keyof typeof TicketPriority]


export const TicketCategory = {
  TECH_SUPPORT: 'TECH_SUPPORT',
  BILLING: 'BILLING',
  SALES: 'SALES',
  OTHER: 'OTHER'
} as const

export type TicketCategory = (typeof TicketCategory)[keyof typeof TicketCategory]


export const TicketChannel = {
  WHATSAPP: 'WHATSAPP',
  EMAIL: 'EMAIL',
  WEB: 'WEB'
} as const

export type TicketChannel = (typeof TicketChannel)[keyof typeof TicketChannel]


export const MessageDirection = {
  INBOUND: 'INBOUND',
  OUTBOUND: 'OUTBOUND',
  INTERNAL_NOTE: 'INTERNAL_NOTE'
} as const

export type MessageDirection = (typeof MessageDirection)[keyof typeof MessageDirection]


export const MessageFrom = {
  CUSTOMER: 'CUSTOMER',
  BOT: 'BOT',
  HUMAN: 'HUMAN'
} as const

export type MessageFrom = (typeof MessageFrom)[keyof typeof MessageFrom]


export const TicketEventType = {
  STATUS_CHANGED: 'STATUS_CHANGED',
  ASSIGNED: 'ASSIGNED',
  TAGGED: 'TAGGED',
  PRIORITY_CHANGED: 'PRIORITY_CHANGED',
  AUTO_REPLY: 'AUTO_REPLY',
  ESCALATED: 'ESCALATED'
} as const

export type TicketEventType = (typeof TicketEventType)[keyof typeof TicketEventType]